refactor(chord-sheet): rename ChordDisplay to ChordSheetViewer, merge ChordContent inline
- Replace ChordDisplay.tsx + ChordContent.tsx with a single ChordSheetViewer.tsx
- Inline ChordContent rendering logic directly into ChordSheetViewer
- Remove ChordContentProps from types.ts and LazyChordContent from lazyComponents.ts
- Update SongViewer and UploadTab imports
- Rename DOM id chord-display to chord-sheet-viewer and update auto-scroll-utils.ts
