+++ frontend/src/components/lazyComponents.ts
@@ -1,7 +1,6 @@
 import { lazy } from 'react';
 
 // Lazy loaded components
-export const LazyChordContent = lazy(() => import('./ChordDisplay/ChordContent'));
 export const LazyStickyControlsBar = lazy(() => import('./ChordDisplay/components/StickyControlsBar'));
 export const LazyChordEdit = lazy(() => import('./ChordDisplay/ChordEdit'));
 export const LazyConfigMenu = lazy(() => import('./ChordDisplay/ConfigMenu'));
